# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OccurrenceResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "string"
          },
          "scheduled_date": {
            "title": "Scheduled Date",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OccurrenceStatus"
          }
        },
        "required": ["id", "routine_id", "scheduled_date", "status"],
        "title": "OccurrenceResponse",
        "type": "object"
      },
      "OccurrenceStatus": {
        "enum": ["pending", "completed", "skipped", "missed"],
        "title": "OccurrenceStatus",
        "type": "string"
      },
...
      "RoutineCreateRequest": {
        "properties": {
          "day_of_month": {
            "anyOf": [
              {
                "maximum": 31.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "days_of_week": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Week"
          },
          "interval_weeks": {
            "default": 1,
            "maximum": 52.0,
            "minimum": 1.0,
            "title": "Interval Weeks",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "schedule_type": {
            "default": "weekly",
            "title": "Schedule Type",
            "type": "string"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["title"],
        "title": "RoutineCreateRequest",
        "type": "object"
      },
      "RoutineResponse": {
        "properties": {
          "day_of_month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "days_of_week": {
            "items": {
              "type": "integer"
            },
            "title": "Days Of Week",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interval_weeks": {
            "title": "Interval Weeks",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "schedule_type": {
            "title": "Schedule Type",
            "type": "string"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "status": {
            "$ref": "#/components/schemas/RoutineStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "notes",
          "schedule_type",
          "days_of_week",
          "day_of_month",
          "interval_weeks",
          "starts_on",
          "status"
        ],
        "title": "RoutineResponse",
        "type": "object"
      },
      "RoutineStatus": {
        "enum": ["active", "paused", "archived"],
        "title": "RoutineStatus",
        "type": "string"
      },
      "RoutineUpdateRequest": {
        "properties": {
          "day_of_month": {
            "anyOf": [
              {
                "maximum": 31.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "days_of_week": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Week"
          },
          "interval_weeks": {
            "anyOf": [
              {
                "maximum": 52.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Weeks"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "schedule_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Type"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "RoutineUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal planning platform API",
    "title": "Planforge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/routines": {
      "get": {
        "operationId": "list_routines_endpoint_api_routines_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoutineResponse"
                  },
                  "title": "Response List Routines Endpoint Api Routines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Routines Endpoint",
        "tags": ["routines"]
      },
      "post": {
        "operationId": "create_routine_endpoint_api_routines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Routine Endpoint",
        "tags": ["routines"]
      }
    },
    "/api/routines/occurrences/{occurrence_id}/complete": {
      "post": {
        "operationId": "complete_occurrence_endpoint_api_routines_occurrences__occurrence_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Occurrence Endpoint",
        "tags": ["routines"]
      }
    },
    "/api/routines/occurrences/{occurrence_id}/skip": {
      "post": {
        "operationId": "skip_occurrence_endpoint_api_routines_occurrences__occurrence_id__skip_post",
        "parameters": [
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skip Occurrence Endpoint",
        "tags": ["routines"]
      }
    },
...
    "/api/routines/{routine_id}": {
      "patch": {
        "operationId": "update_routine_endpoint_api_routines__routine_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Routine Endpoint",
        "tags": ["routines"]
      }
    },
    "/api/routines/{routine_id}/archive": {
      "post": {
        "operationId": "archive_routine_endpoint_api_routines__routine_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Routine Endpoint",
        "tags": ["routines"]
      }
    },
    "/api/routines/{routine_id}/pause": {
      "post": {
        "operationId": "pause_routine_endpoint_api_routines__routine_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Routine Endpoint",
        "tags": ["routines"]
      }
    },
    "/api/routines/{routine_id}/resume": {
      "post": {
        "operationId": "resume_routine_endpoint_api_routines__routine_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Routine Endpoint",
        "tags": ["routines"]
      }
    },
...
  }
}